Перенос слайдера и добавление Enum вместо координат из тест. данных
